Adding custom hooks for data fetching
