+++ src/components/Routes/Routes.test.jsx
@@ -4,7 +4,7 @@ import { createMemoryHistory } from 'history';
 import { Router } from 'react-router-dom';
 import Routes from './Routes.component';
 import Theme from '../App/App.styled';
-import { SearchProvider } from '../../Hooks/searchContext';
+import { SearchProvider } from '../../Hooks/SearchProvider';
 
 describe('App Component Tests', () => {
   it('Should redirect to home', async () => {
